delete product

+++ Server/controllers/product.js
@@ -1,6 +1,12 @@
 const prisma = require("../config/prisma")
 const cloudinary = require('cloudinary').v2
 
+cloudinary.config({ 
+    cloud_name: process.env.CLOUDINARY_CLOUND_NAME,
+    api_key: process.env.CLOUDINARY_API_KEY, 
+    api_secret: process.env.CLOUDINARY_API_SECRET
+});
+
 exports.create = async(req, res) => {
     try {
         const { title, description, price, quantity, categoryId, images } = req.body
@@ -110,7 +116,33 @@ exports.update = async(req, res) => {
 exports.remove = async(req, res) => {
     try {
         const { id } = req.params
+        // include image
+        const product = await prisma.product.findFirst({
+            where: {
+                id:  Number(id)
+            },
+            include: { images: true}
+        })
 
+        if(!product){
+            return res.status(400).json({ message: 'Product not found'})
+        }
+
+        console.log('z', product)
+        
+        // promise image
+        const deleteImage = product.images
+            .map((image) => 
+            new Promise((resolve, reject) => {
+                cloudinary.uploader.destroy(image.public_id, (error, result) => {
+                    if(error) reject(error)
+                    else resolve(result)
+                })
+            })
+        )
+        await Promise.all(deleteImage)
+
+        // delete producr
         await prisma.product.delete({
             where: {
                 id: Number(id)
@@ -232,11 +264,7 @@ exports.searchFilters = async(req, res) => {
     }
 }
 
-cloudinary.config({ 
-    cloud_name: process.env.CLOUDINARY_CLOUND_NAME,
-    api_key: process.env.CLOUDINARY_API_KEY, 
-    api_secret: process.env.CLOUDINARY_API_SECRET
-});
+
 
 exports.createImage = async(req, res) => {
     try {
